feat(core): moved bodyParser$ and logger$ into separate packages
