+++ packages/core/src/index.ts
@@ -1,8 +1,6 @@
 export * from './effects/effects.interface';
 export { combineRoutes } from './effects/effects.combiner';
-export * from './middlewares/bodyParser.middleware';
 export * from './middlewares/error.middleware';
-export * from './middlewares/logger.middleware';
 export * from './operators';
 export * from './http.interface';
 export * from './http.listener';
