+++ packages/middleware-body/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@marblejs/core",
+  "name": "@marblejs/middleware-body",
   "version": "0.2.2",
-  "description": "Reactive Node APIs made easy",
+  "description": "Body parser middleware for Marble.js",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "scripts": {
@@ -29,13 +29,8 @@
   },
   "homepage": "https://github.com/marblejs/marble#readme",
   "peerDependencies": {
+    "@marblejs/core": "^0.2.2",
     "rxjs": "~6.1.0"
-  },
-  "dependencies": {
-    "path-to-regexp": "^2.2.1"
-  },
-  "devDependencies": {
-    "@types/path-to-regexp": "^1.7.0"
   },
   "publishConfig": {
     "access": "public"
