+++ packages/middleware-body/tsconfig.json
@@ -1,0 +1,21 @@
+{
+  "extends": "../../tsconfig.json",
+
+  "compilerOptions": {
+    "outDir": "./dist/",
+    "declaration": true,
+    "typeRoots": [
+      "./node_modules/@types",
+      "../../node_modules/@types"
+    ]
+  },
+
+  "inputs": [
+    "src"
+  ],
+
+  "include": [
+    "node_modules/@types",
+    "src/**/*.ts"
+  ],
+}
