+++ packages/middleware-logger/src/index.ts
@@ -1,7 +1,6 @@
+import { Effect, HttpRequest } from '@marblejs/core';
 import chalk from 'chalk';
 import { tap } from 'rxjs/operators';
-import { Effect } from '../effects/effects.interface';
-import { HttpRequest } from '../http.interface';
 
 const getStatusCode = (statusCode: number): string =>
   statusCode >= 400
@@ -13,20 +12,27 @@ const getEndTime = (hrStartTime: [number, number]): number => {
   return Math.round(hrEndTime * 100) / 100;
 };
 
-const printLog = (method: string, url: string, statusCode: string, endTime: number) => {
+const printLog = (
+  method: string,
+  url: string,
+  statusCode: string,
+  endTime: number
+) => {
   console.info(`%s %s %s %dms`, method, url, statusCode, endTime);
 };
 
-export const logger$: Effect<HttpRequest> = (request$, response) => request$
-  .pipe(
+export const logger$: Effect<HttpRequest> = (request$, response) =>
+  request$.pipe(
     tap(request => {
       const startTime = process.hrtime();
 
-      response.on('finish', () => printLog(
-        request.method!,
-        request.url!,
-        getStatusCode(response.statusCode),
-        getEndTime(startTime),
-      ));
-    }),
+      response.on('finish', () =>
+        printLog(
+          request.method!,
+          request.url!,
+          getStatusCode(response.statusCode),
+          getEndTime(startTime)
+        )
+      );
+    })
   );
